Add files via upload
Initial commit of chromedriver image gathering for neural net training.  Only gets images up to "show more" - still needs the "show more" link clicked.  Also, this does not include the needed chromedriver Selenium binary, which is a dependency (and the coder has to enter the proper path to that file).