Add files via upload
This is not finished yet - I am rewriting the javascript neural net (in this folder), but am going to use numpy matrix multiplication instead of double for loops in the forward and back propagation.  Also, I need to integrate the work I did for image gathering + image conversion into this.